gpio-mxc: fix issue multiply setting pads wakeup on the same gpio bank
This patch fixes the issue in the case of we need to configure more gpio lines
to pad wakeup on the same gpio bank. Without the fix only the first pad configured
as pad wakeup works fine. For the others gpio lines the below error appears:
  - gpio-mxc xxxxxxxx.gpio: set pad wakeup failed, ret -17

Example of device tree with multiply pad wakeup on the same gpio bank:
&lsio_gpio5 {
	/* Enable wakeup-source from all gpio-keys*/
	/* total pad wakeup number in gpio5 */
	pad-wakeup-num = <4>;
	pad-wakeup = <IMX8QM_USDHC1_DATA4 4 19>, <IMX8QM_USDHC1_STROBE 4 23>, <IMX8QM_USDHC1_DATA5 4 20>, <IMX8QM_USDHC1_DATA6 4 21>;
};