drivers: net: fec: Add phy-reset-gpios-fec2 property
If fec1 reset line is asserted when fec0 is probed,
fec1 will fail to detect the phy id. When fec0 and fec1
share the same mdio, this property allows the fec0
initialization to also control the reset line for fec1.

According to the comment in fec_enet_mii_init:

> The i.MX28 dual fec interfaces are not equal.
> Here are the differences:
>  - fec0 supports MII & RMII modes while fec1 only supports RMII
>  - fec0 acts as the 1588 time master while fec1 is slave
>  - external phys can only be configured by fec0
>
> That is to say fec1 can not work independently. It only works
> when fec0 is working. The reason behind this design is that the
> second interface is added primarily for Switch mode.
>
> Because of the last point above, both phys are attached on fec0
> mdio interface in board design, and need to be configured by
> fec0 mii_bus.

When fec0 is probed, fec_enet_mii_init will call
of_mdiobus_register which will search the mdio bus for phys.
Due to the explaination above, it will not call it again
when fec1 is probed.

If the fec1 reset line is held low while fec0 is probed,
it's phy id will not be detected and it will fail to probe.

On Symphony board 1.6, the phy reset line does not have a
pullup resistor. This patch provides a workaround allowing
fec0 to release the reset line on the fec1 by providing the
reset gpio via phy-reset-gpios-fec2.

An alternative workaround is to hardcode the phy id using the
compatible string. For example, for the adin1300:
compatible = "ethernet-phy-id0283.bc30";
This is not an acceptable workaround because different versions
of the Variscite Symphony board support the ar8033 and adin1300
on the same address.

Signed-off-by: Nate Drude <nate.d@variscite.com>